SpecialBlock [Codex]: use blank cells for no data in tables
A blank cell is the standard way we represent 'no data' in data tables
in MediaWiki. Screen readers understand this without the need for an
ARIA label. Additionally, these data tables consume a lot of UI real
estate at Special:Block, so it's better to avoid unnecessary messages.

Other changes:
* Simplify util.formatTimestamp() to take only a single argument
* In the active blocks table, have the timestamp link to
  Special:BlockList with the block ID, consistent with other places we
  list active blocks.
* Fix bug with checking if block parameters are present
* Use mw.util.isInfinity() instead of hard-coding infinity values (which
  can differ depending on the API used)
* Fix a few Vue linting warnings

Bug: T376746
Change-Id: I797edf21e63959a91a09e9771b7e231a6137c57c