StaticArrayWriter: Optimize expensive logic for speed
This little library is used by our LocalisationCache (LC). The
output is heavily cached and this code ends being called very rarely.
But when it is called it's surprisingly expensive.

This is relevant especially in test scenarios where caches are not
primed or intentionally purged.

With this patch the relative time spend in this class alone goes down
from 24% to 15%, according to XDebug. The total time spend in
LocalisationCache goes down from 37% to 28%. I find this a
surprisingly big difference considering the code still executes the
same deeply nested loops as before.

Reasons are:
* str_repeat is entirely gone.
* Single-use methods are inlined, avoiding PHP's expensive call stack
  overhead.
* Instead of checking if var_dump returns the string "NULL" we check
  the original value.

The tests prove that this patch doesn't cause any visible change.

To test this locally you can add this to your LocalSettings:
  $wgLocalisationCacheConf['forceRecache'] = true;

Bug: T381895
Change-Id: I3befa7f0215980d423840a65bf9e79b4b3c5bb1c