tests: Move code in resetNonServiceCaches() out of a loop
This appears to be a mistake, or is it? The makeJobQueueGroup() is a
factory method with an instance cache and always returns the same
instance anyway.

In an example PHPUnit test run I did the number of calls to the
makeJobQueueGroup() method goes down from 14,000 to 380. The total
time spend in resetNonServiceCaches() goes down from 14% to 9%,
according to XDebug. I find this surprisingly significant.

The issue apparently exists since 2018, see Ia4b7871. It might
be that the original JobQueueGroup::singleton() call was cheaper
back then. The replacement that was added in 2021 via Ic870911 is
surprisingly expensive.

This is motivated by some of the discussions in T225730.

Change-Id: I910023a01f8a5d60bab0b4c45a4f742a51616c45