jobqueue: Re-arrange makeJobQueueGroup for performance
This simple re-arrangement of existing code actually cuts the time
spend in this method in half. Since I put this patch on top of
I910023a it doesn't have a big practical impact any more. We would
get much more impressive results when we test this without the
previous patch I910023a. Still I think it's worth it. I mean, why
execute the same, possibly expensive code twice? And why execute it
at all if the result isn't going to be used?

This is motivated by the discussions in T225730 and mostly relevant
in test scenarios.

The reason for this code being more expensive than expected can be
seen in Ic2293c4 where a bunch of code was moved out of the
isset( $this->instances[…] ) check, as well as a little bit of
duplication introduced. This duplication appears to be cheap but
unfortunatelly isn't. It especially adds up when executed thousands
of times.

Change-Id: Ic3b50f5998a9c939ed5779ccf6a2f8830cc86809