parser: Handle invalid titles gracefully in parser functions
Direct construction of TitleValue is not ideal for untrusted text.
However, we cannot use TitleParser::makeTitleValueSafe here because
it will fully parse anything that looks like interwiki prefix inside
the title text and we don't want that for interwiki and interlanguage
parser functions.

For instance in `{{#interwikilink:Project|Test:Link}}`, we don't want
to parse 'Test' as interwiki prefix, but we want to ensure the entire
'Test:Link' string to be in valid title text form.

Bug: T381977
Change-Id: I9cdde3cfbd4797128fb5596d5b5f7b69467488f2