maintenance: Factor out a few methods from SchemaMaintenance
SchemaMaintenance contains non-trivial logic to generate a standard SQL
comment, and to reformat the autogenerated SQL. Being inside a
Maintenance class, it is unpleasant to test because of all the
dependencies on CLI arguments etc.

So, extract a new SchemaGenerator class that is responsible for
validating a JSON file, and autogenerating a canonical .sql file from
it. This belongs to the maintenance realm (as opposed to the DBMS lib)
due to specific assumptions around formatting, as well as the script
name and instructions in the comment.

Errors are passed around using the existing
AbstractSchemaValidationError, so we don't have to rely on
Maintenance::fatalError.

As part of the refactoring, update the JSON path canonicalisation done
before outputting the SQL comment, so that the path is always relative
to the extension or core, regardless of where the script is run, to
ensure consistency and make it easier to test.

Adjust tests, duplicating some test cases as needed.

Bug: T381981
Change-Id: I358cc24798f518d7c3dc03c60bcc0702af252fee