db: Drop AbstractSchemaValidator::$missingDepCallback
For starters, the property is really only used by checkDependencies(),
which in turn is only called by AbstractSchemaValidationTest. Everywhere
else, all this code was a no-op.

But for PHPUnit tests, we must not tolerate missing composer
dependencies. In fact, our PHPUnit bootstrap even runs the
CheckComposerLockUpToDate script and exits immediately if dependencies
are not up-to-date, let alone missing.

Instead, move the dependency checks to the constructor, so that they're
always executed, and throw an exception when they fail, instead of
relying on the callback. In practice:
- For PHPUnit tests there's no real difference thanks to the
  composer.lock check mentioned above.
- For maintenance scripts, when dependencies are not available, we now
  fail with a useful exception, instead of a generic "Class X does not
  exist" message.

Bug: T381981
Change-Id: I893dc6a52262f8a63fce3ec5eecdd91dd9f339a5