SpecialBlock [Codex]: improve handling of clearing of the target user
CdxLookup doesn't fire the 'change' or 'input' events when the clear
button is clicked. Instead we need to add a specific listener for it.

onClear() is added which clears the targetUser in the store, and calls
$reset() since this is effectively what the user is going for.

In order to ensure success messages are cleared post-block, we track
'success' in the store, and clear it in $reset().

Also remove the placeholder blank space that was added as part of
I0e6efc53cf. This isn't needed now that we hide the form entirely when
the target is cleared. This way, there's no odd space between the target
input and the error message associated with it (since we wouldn't show
convenience links for an invalid target, anyway).

Finally, set focus on the input field when clearing and when initially
mounted.

Bug: T380968
Change-Id: I37013ba8127de6cd440db7a205bda9373e6823a1