SpecialBlock [Codex]: fix reblocking and rework store
Attempt to give the store some structure, somewhat mimicking how Option
Stores are structured in having a section for state prop (refs), getters
(computed props), and actions (functions). Add JSDoc blocks for all
exported symbols, detailing their purpose and reactivity. Remove
anything from the store that isn't shared across components or necessary
for other store logic, and remove ref()s from constants.

$reset() is now resetForm(), since we're resetting to default values (as
established in SpecialBlock.php) rather than the initial state of the
store. We call this when adding a new block, or when resetting the form
via the clear button in the UserLookup component. As such, the initial
values for properties are restored to the ref declarations.

Add basic selenium test that creates an account, logs in as an admin,
blocks the newly created user, and asserts the success message. This
test passes an example expiry via URL query parameter, effectively
testing that it gets set properly in SpecialBlock.vue.

Bug fixes:
* Ensure alreadyBlocked prop is set when it should be,
  and reblock=1 is passed to API
* Pre-fill form fields form URL parameters, as applicable
* The default value for the wpCreateAccount URL param should be true
* Hide the form when the targetUser changes
* Fix Vue warning in SpecialBlock.vue

Bug: T382035
Change-Id: I864e7cd259b14fbf6444b4c54eeedd6fb01db844