Re-generate all schema change files
This gives us a solid baseline to use for automated tests. Whitespace
differences will be ignored by the test, but regenerate these anyway
because they're much more readable now, and it's generally nice to have
a clean baseline.

Differences are due to:
- JSON files being moved in T382030
- JSON paths in the autogenerated comment are now standardized
  (I358cc24798f518d7)
- Whitespace changes from Ib2f60b492ced8ff
- Actual SQL changes from newer versions of doctrine/dbal

There's also one important exception: the SQLite schema change for
patch-page-rename-name_title-index no longer matches the JSON file since
I17e73d5ef1. There doesn't seem to be a super-simple way around this,
because we don't know the state of the database pre-update. On top of
that, I believe that removing the field from the .sql file as done in
I17e73d5ef1 could've caused data loss when upgrading from old versions:
the column could be silently dropped before running the DB update that
correctly migrates and drop the column. This is something to keep in
mind going forwards, and T381981 will prevent it from happening that
easily. But for the update in question, it's probably too late, so just
update the JSON file to match the .sql files.

Bug: T381981
Bug: T382030
Change-Id: I3a3acc0935409d4601eb172664580b13d5fe2340