SpecialBlock: Handle expiry datetime in JS the same as in PHP
Change the regex in JS to look only for a match at the beginning
of the expiry datetime string for Y-M-DTh:m. That's what's already
being done in PHP.

This doesn't fix the bug with timezones, e.g. a URL prefill such
as Special:Block?wpExpiry=2024-12-16T10:00:00+0800 should result
in an expiry time of 02:00 UTC; this will be fixed separately.

Bug: T380963
Change-Id: I03c8cb50ee1b49c66c64d62c6e754199d6b49526