SpecialBlock: Handle block ID or 'newblock' when saving form
When multiblocks is enabled, send either the block ID or the
'newblock' parameter to the API when saving or creating a block.

Also fix the form's button text to be correct when creating a
new multiblock.

This moves the enableMultiblocks flag into the store, and uses
in (some of) the various places it was being checked.

Bug: T380187
Change-Id: Ie54540888aa05a6dc905512561a4cefd98c9d2f6