Revision: make RevisionRecord::getPage() return a value object
RevisionRecord::getPage() returns a PageIdentity, which may
be a Title or a WikiPage. In the context of an ongoing page deletion,
the state of these objects may change and the page ID may get set to 0.
To provide a reliable stable interface that can be used e.g. in a
deferred update or job, we need to return an immutable value object
instead.

NOTE: this may break callers of getPage() that downcast to Title,
or that rely on the page ID contained in the returned PageIdentity
to be updated dynamically during page creation or deletion.

This is a modified version of I7440f8f30b05949. It reverts I4ed5406a8b3aa9, which reverted the original patch. The main difference
to the original patch is the additional check in the constructor of the
RevisionRecord base class and the regression test in SignatureValidatorTest.

Bug: T380536
Change-Id: I0ea915983b6187090f46ad23c9b582196b4be85f