SpecialBlock: Change hideName to hideUser to match field name
Although the user right, block attribute, and a system message all
use the name 'hidename', the form field and other front-end variables
are called 'hideuser', so this updates a few places to match the
latter. This makes it easier to understand, and also fixes a bug
where it was being referred to by the wrong name.

Bug: T382302
Change-Id: Icfd1738c72ff8f03b68f4976d3b9419440841fdb