maintenance: Fix incorrect error message for missing short value
The logic bug pre-existed, but the concrete issue here was introduced
last year in Ib67667fea (a9f47fd5ad), which changed "$option parameter"
to "--$option parameter" for both long and short option messages.

Refer to the option given rather than the mapped name.
Like I7ab1206e3, this is also due to a re-used variable name where
$option is changed in order to look up the metadata about the option,
but this was then re-used for the message output.

It is fairly okay in the common case of `--value (-v)` where --value
if you used -v actually, is at least something you can find in the
help. But, in the case of an option that only has a short name,
it was mis-calling the short option as a long option.

Either way it's an improvement to refer to what the user actually
did. Before:

```
$ php maintenance/eval.php -d

ERROR: Option --d needs a value after it!
```

Fixed:

```
$ php maintenance/eval.php -d

ERROR: Option -d needs a value after it!
```

Change-Id: I3419a831e71c63bc64ba3d7f517fb2cd00707136