objectcache: Tag SerializedValueContainer as `@internal`
There is no reason for code outside the BagOStuff component to use
this class. It is used solely for encoding, detecting, and decoding
internal value wrappers. Any value passed to or returned by BagOStuff
methods does not involve this class.

Between MW 1.34 and 1.43 this class had more public methods, which
have since been removed.

As an internal class, move this freely into the flat list of
objectcache/ classes.

Bug: T353529
Bug: T364652
Change-Id: I22178bdaf6c9762a32e7428d3a286583ca99bfcf