UserSelectQueryBuilder: Do not consider unregistered actors to be named
Prior to this patch, UserSelectQueryBuilder::named was
implemented as an equivalent of `!User::isTemp()`. Unfortunately,
that is not fully correct. An IP address is a non-temporary actor,
but it is not a named actor. Actors must be BOTH non-temporary
AND registered to be considered named.

This patch fixes the bug.

Bug: T382889
Change-Id: Ie24ccaa0c446150f61772ecfbebe186757fdf26e