NoopSpan: optionally activate
Why:
If tracing is enabled, but the sampling probability is <100%, we may
set a NoopSpan as the root span.  Without this patch, the NoopSpan
cannot be activated, and thus creating child spans will fail with
PreconditionException.

This is a minimal patch to allow the NoopSpan to be activated, while
not affecting existing testing codepaths that don't require a root span.

Bug: T340552
Change-Id: Ib8a2d7b483190bba21a2f3b82fcd0e0b5756e750