filebackend: Improve debug messages by passing native context or CSV
By passing arrays as native PSR3 context, the values end up as part
of the structured debug logs in Logstash where they can be shown
on-screen, selected, and copied into bug reports as their original
values instead of with JSON-escaping for various characters.

This is not a big deal today since FormatJson defaults to
JSON_UNESCAPED_SLASHES (which is important because most of these
values contain file paths that definitely include slashes). But,
this will change as part of T382910 by removing use FormatJson, and
inlining these gets tedious so instead pass it as native data where
the data is non-trivial, and for others embed the paths within the
string unchanged comma-separated.

Bug: T382910
Change-Id: I79fba128a8129b9ac4da17b513b5da0b03d90bf9