htmlform: Add HTMLOrderedMultiselectField for use in HTMLForm
This field uses a new widget based on OO.ui.MenuTagMultiselectWidget to
allow choosing from multiple options in a dropdown. It internally uses
a textarea ensuring that order of selected chips are preserved. No-JS
fallback is a textarea with no dropdown, but validations ensure that
arbitrary values entered are not accepted.

Bug: T382995
Change-Id: I4c8548a820c6035b28efdb47df8d7cbeeb407031