Make onConditionalDefaultOptionsAddCondition call easier to find
Previously, ServiceWiring.php called the hook directly using
the HookContainer. This was making the usage of the hook
difficult to find, as eg. PHPStorm would claim
onConditionalDefaultOptionsAddCondition has zero usages.

This patch changes the call to make use of the HookRunner,
thus being safer (thanks to typehints) and easier to find
(thanks to actually calling a method).

Change-Id: I5e515301f0664fedf110d401868262143d62b63c