file: Acquire the file id in all cases
This used to be called getFileIdFromName() because acquire…() was called
a couple lines above, except that the line is not called for new uploads
(only for re-uploads). This caused new files have their first revision
holding fr_file as zero in the beta cluster (end of the world, I know).

The overhead of acquire even if the row exists is negligible (it calls
get first and if it doesn't exist, it inserts it) so let's just call
it.

Bug: T383093
Change-Id: I47fee73b4510b7aa1858c5c607bb626ec7032645