Stats: clarify TimingMetric::observe() units and discourage in new code
In the previous implementation with statsd/Graphite, the actual unit of
what was value did, defacto, not matter. Seconds and even bytes worked
just fine. However, with Prometheus it is essential to store timings
with the correct units, because the buckets for the histogram are
based on absolute thresholds.

While there exist `observeSeconds` and `observeNanoseconds` helper
methods on this class, the main method promoted today is `observe()`
which was missing essential information in its name or doc about what
unit to provide. This is made worse by the fact that all documentation
and migration instructions emphasize that the metric should be named
with a `_seconds` suffix, and reference Prometheus documentation which
explicitly states that SI units like seconds should be stored.

This disconnect caused T383208 and will keep being a source of
confusion and bugs as long as this method is promoted.

Ideally, this would have been part of Ieb24a5f723 (commit 37d29e8),
but second best time to fix this is now.

Bug: T364240
Change-Id: Ie92fa7897085bef55b37b02ee4ddba774dc2ed9c