specials: Reduce calls to permission checks in PagerTools
Authority::probablyCan needs to load page restriction data for each
different title it is called for. The data is loaded for each title in a
separate query. (RestrictionStore stores this for 1 day in a WanCache.)
Reduce the amount of calls to Authority::probablyCan by checking the
cheaper conditions in outer if.

On Special:Contributions this avoids to load page restriction data for
non-top rows.

Bug: T378764
Change-Id: I5bebb35163b04cb39beb0ba728bf4324a120f6d6