user: Use TS_MW to compare timestamps in UserRegistrationProvider
It is valid to use TS_MW for chronological compare as the timestamp
stored in a string is ordered from the most significant datetime
information (year) to the lowest (seconds).
It is not needed to convert to unix.
It would not be valid to use TS_MW format to calculate time differents.

This is how it is used in the database schema from mediawiki as well.
Also replace array + min() with variable + if.

Change-Id: Id41e660f3952876e20ac4e88502d7e2757299573