user: Use TS_MW to compare timestamps in ConditionalDefaultsLookup
It is valid to use TS_MW for chronological compare as the timestamp
stored in a string is ordered from the most significant datetime
information (year) to the lowest (seconds).
It is not needed to convert to unix.
It would not be valid to use TS_MW format to calculate time differents.

This is how it is used in the database schema from mediawiki as well.

Bug: T378764
Change-Id: Ifc8dbc0a87d286fb63f200b1f255cb11cc70b1f3