installer: Simplify the information box
The image passed to ::getInfobox is never passed to ::infoBox() it is
only called twice with an image which also doesn't work, it is never
called with 'class' option so remove both.

That also means the img element is fully useless and is causing
user-facing errors, so just remove it.

Bug: T383646
Change-Id: I4abe9f608e9b5dd657e3948988f28ad991495f25