phpunit: Don't override --bootstrap if supplied
SemanticMediaWiki creates its own bootstrap and displays
a bunch of information. In [0] this broke this.

We already look for bootstrap anyways, we just don't use it.

We fix this by checking for bootstrap and only setting a default
if not supplied.

[0] d2a30096f142a582ede952c28d2edd83e5bf1598

Change-Id: Iebb1949cc6a253640f40f1ac048ab50f99cac02a
(cherry picked from commit d43c69b25f05801ac3c255a2dfdbe03e820029ca)