Storage,MessageCache: Reference TTL constants from BagOStuff
Mark `ExpirationAwareness` as internal, as it is effectively a private
trait for the benefit of de-duplication between MapCacheLRU,
BagOStuff, and WANObjectCache APIs.

Callers should reference constants from the object ($cache::TTL_DAY)
or type (BagOStuff/WANObjectCache) that their code is interfacing
with. Fix a handful of cases that didn't do this.

Bug: T353529
Change-Id: Ic46e09c848e69db3d22219d2231274c81c0cb275