SpecialBlock [Codex]: hide all of CodexHTMLForm except the target
This patch drops no-JS support for the Codex version of Special:Block.
It can later be added when Codex PHP is available. All we really need is
the target field, so show just that and with the module styles to match.

If a target is provided via URL parameter, have the field be disabled
until the Vue app loads.

This patch also removes OOUI completely from the Codex version. The
error messages are instead created with Html::errorBox().

Bug: T377529
Change-Id: Ica7cee99f09f8abaafd28d67f0baa254b91bc41e