SpecialBlock: Handle URL param remove=1 to open dialog
Delete the showBlockForm variable and use store.formVisible directly,
to simplify things. Also make blockId non-reactive because it never
gets updated (we update store.blockId and having another var with a
block ID might be confusing).

Then we use the existing system of checking for the URL-provided
block ID's data, and either carrying on as normal to open the
editing form, or instead opening the removal dialog.

Not done in this is any update to the URL, so after removing a block
subsequent reloads of the page result in an error.

Bug: T383746
Change-Id: I251f93ab3db937cee5a1cfb93094680ae23f841f