session: Support REST API in Session::makeException()
Allows nicely formatted error messages when the session provider
fails to authenticate a REST API request.

This is not ideal in that (unlike ApiBeforeMain) RestCheckCanExecute
is only called if a module is identified, so for routes not matching
any module prefix the error message will be different, but adding
another hook just because of that did not seem worth it. On private
wikis the different error message would give some clue on what
extensions are installed, but the REST API doesn't really try to hide
that information from unauthorized users anyway.

Bug: T252591
Change-Id: Ib7cb7bc38234113f242c8c3349a21e0649440691