ComposerLaunchParallel: Configure split group count from environment
Defines a DEFAULT_SPLIT_GROUP_COUNT of 8, matching current behavior.
Allows this to be overridden with an environment variable,
QUIBBLE_PHPUNIT_PARALLEL_GROUP_COUNT.

Value must be an integer greater than 2. Throws
InvalidSplitGroupCountException for invalid counts.

This is handled in a static getSplitGroupCount(), since it needs to be
available from the static launchTests().

I tested this in mediawiki-docker by running something like:

  $ composer phpunit:prepare-parallel:default
  $ composer phpunit:prepare-parallel:split-file tests-list-default.xml

Then to check the default value:

  $ composer phpunit:parallel:databaseless

To check a non-default value:

  $ QUIBBLE_PHPUNIT_PARALLEL_GROUP_COUNT=4 composer phpunit:parallel:databaseless

To check for an exception on an invalid value:

  $ QUIBBLE_PHPUNIT_PARALLEL_GROUP_COUNT=foo composer phpunit:parallel:databaseless

I'm not sure it _works_, but it doesn't blow up.

Bug: T383254
Bug: T383389
Change-Id: I20dd9663ac0dccd974ae84b2b89d13a3f2842f9c