SpecialBlock [Codex]: add codex.userlookup hook for custom components
Extension:CheckUser needs to add a "Show IP" button in the interface. To
keep extension-awareness out of Core, we add a 'codex.userloopup' JS
hook that provides a shallowRef of an array of components. A sepearate
patch to CheckUser will make use of it. For now, we provide custom
components only the targetUser as a prop.

The hook name is intentionally not tied to Special:Block because the
UserLookup component is slated to be a shared MW component, where
presumably the same CheckUser integration would be desired.

Bug: T383663
Change-Id: If061f465cf08d3348337103858e620dbd1cca219