mediawiki.api: Allow setting Api-User-Agent via mw.Api constructor
Allows passing a user agent in the constructor, which will be sent
in the Api-User-Agent header (changing the User-Agent header is
not possible under the browser security policy, and it also holds
its own valuable information).

Setting Api-User-Agent was possible already via the `headers` option,
but was uglier and less obvious. Also, this patch adds a default of
"MediaWiki-JS/<version>"

* different from the MediaWiki PHP user agent ("MediaWiki/<version>"),
* clearly related to MediaWiki frontend JavaScript
* avoid generic words like "mw.Api" which are easily mistaken for
  any number of "MediaWiki" "API" things.

Usage:

    const api = new mw.Api( { userAgent: 'MyGadget/1.0.1' } );

This allows gadgets, user scripts, MediaWiki core, and extension code
to easily identify which client makes an API request.

Identifying the client helps:
* WMF and other sysadmins, to debug and find the source of a problem.
* MediaWiki maintainers, to assess what disruption would be caused
  by backwards-incompatible changes in the API response or with
  proactively fixing clients before such changes.
* Gadget and extension maintainers, by enabling them to use
  Special:ApiFeatureUsage to check whether they rely on deprecated
  APIs.

Bug: T373874
Change-Id: I08cbe95675c7e3b6b3a20875f3d061afdc04efd6