SpecialBlock: Scroll to erroneous field rather than just the error
Scroll up to the .cdx-field that is the parent of the error message,
if there is one. Because the error messages are shown below the
field, without this it's possible to be scrolled up to see only
the error message and then have to scroll up more to see what field
it actually applies to.

Bug: T383975
Change-Id: I126f064e19d37e1e858df2821205f914fbdec855