Suppress error log for empty headings
It's a common enough occurrence that heading content will normalize to
an empty string, which is set as the heading id when storing section
metadata.  However, Parsoid will determine that to be an invalid id when
storing node data and reassign it a Parsoid generated id.  Thus, section
mapping will never select the heading node.  There's no need to generate
an error log for this case.

Parsoid could eagerly assign a generated id when it runs into this
situation but that wouldn't match the legacy parser.

T368722 suggests either considering empty headings as invalid and
rendering the markup syntax or emitting a lint to have the uses cleaned
up.

There's a test demonstrating empty headings section wrapping  in
I6aea9351786264bd79dddd7f5b234c923ef172a6

Bug: T375002
Bug: T368722
Change-Id: Ibf678b343dc6e7db6eee3ffd687ca2cbd72b7250