RevisionStore: Query by page_id when available
Historically these methods required a LinkTarget. They have been
expanded to also accept a PageIdentity via Ica8da21. Providing a
LinkTarget was (soft-) deprecated the same time. That was in 2021,
version 1.36.

This patch gets rid of the extra indirection of converting
LinkTarget to Title first, but does the query directly by page_id.
We expect this to be more stable and potentially faster.

This got motivated by the WIT team's work on T352019.

Bug: T272886
Bug: T380677
Change-Id: Id79dcd26821af2cf2ec8cc346fb7ba6863b53466