Revisions: Rewrite a RevisionRecordTests for readability
I find the code quite surprising and hard to read because it assumes
that a variable will continue to live outside of the scope of the
"catch". This might work in PHP but not necessarily in other
languages and can be confusing because of this.

PHPUnit does have standard utilities to work with exceptions:
* In the first example we don't need to catch the exception just to
  make the test fail. PHPUnit already does this.
* In the other two examples the PHPUnit way to do this is with
  expectException.

The extra assertions in the end don't do much but are needed to not
have the tests marked as risky.

Change-Id: I62220a93ac176c42d4a44724484d05db89fb601c