SpecialBlock: Scroll to the error's fieldset instead of field
Scroll up to the parent fieldset of any error, to ensure that
more context is visible to the user. The expiry field's label is
not within its field (because the label is with the radio button)
so this makes sure that more than just the input box is made visible.

Bug: T383975
Change-Id: I77f1c7dbe6ea68244e35b8842d6dbdf20e6860f0