tests: Fix stale setup/teardown tracking
Previously, if the setupLevel ever exceeded 1 due to one faulty test
tearDown(), it would never reach 0 again and would also throw an
exception in every subsequent test, which was quite confusing. Fix that
by resetting the tracking before throwing the exception; and also turn
the “level” into a boolean flag, since I don’t see any need to track it
as an int.

Bug: T384588
Change-Id: Ic24dda3953e703151bda1c99c098f1f37abca7f8