Skins: Remove SpecialPages link from toolbox, add to navbar by default
Make the link be in the Navigation sidebar (as the last link) by default by changing the default "sidebar" interface message and adding a new "specialpages-url" interface message. If an existing navigation does not contain the "n-specialpages" item, we append it to the end of navigation. As mentioned in the task, this is done to line up with the other links in the main menu (navigation) vs the toolbox.
As that sidebar is specified through the "sidebar" interface message, the only way to add the link to the main menu without setting great new precedent is to change the default interface message.
In accordance with existing links in the default sidebar, the link to "Special pages" was specified through a URL interface message formatted the same way as the "Recent changes" entry.
The "createSidebarItem" function executes code shared between the addition of n-specialpages and the addition of other sidebar items. Hopefully I got the naming part and the getting-config-and-services part right.
This approach requires some user notice, as users should update their MediaWiki:Sidebar to include n-specialpages. I also wonder about the performance footprint, especially that of foreaching the sidebar to find n-specialpages and that of copying over config, services, etc. every time a sidebar item is generated.

Bug: T333211
Change-Id: I03f818d414c605a3cc87dfd108a8860a09d41d5f