Avoid constructing Title from LinkTarget where not needed
I just find this combination of method calls confusing. A revision
can not be addressed via a LinkTarget. We can only end having a
revision when we started with a proper PageIdentity. This is actually
hard-coded in the RevisionRecord constructor. So why not use that
PageIdentity when we have the guarantee that it's available?

I believe the proposed replacement is also faster. Not only because
we avoid calling some extra code. The new code makes sure that the
page ID is never lost and doesn't need to be queried again.

Change-Id: I3ea3dd8b52566cc9883f3328c7748584fb3d05ff