mediawiki.api: Hoist local 'defaultOptions' variable
The pattern was originally introduced by me I2d4f8b99c (1917ed656c),
because at the time, the doc block for defaultOptions contained useful
information to the public API, and the only way to refer to a static
private member of a class, was after the class itself is defined.

Then, in in Ia9320213e (e365e22ab7) the doc block was converted from
JSDuck to JSDoc with the object given its own top-level imaginary
typedef, which has the benefit of being resolved a bit later rather
than at parse time, thus allowing it to be defined lower in the code
but referenced in the class block. While that change removed the
reason for the pattern, it left the actual pattern behind.

Then, Id1ec60db0 (e352a3a878), it was converted from var to let,
which ossified the supposed non-const nature making it seem
intentional.

Change-Id: Ia70dd6f2dd1246beb0c471dce92b17f9e0233383