mediawiki.api: Remove console warning for legacy token type
The following warnings are logged to the browser console (and e.g.
on the CLI from `npm run qunit` in MediaWiki):

> Use of the "options" token is deprecated. Use "csrf" instead.
> Use of the "email" token is deprecated. Use "csrf" instead.

This is contributing to warning fatigue in CI, with other stuff
slipping and being missed because it's normal.

We could call suppressWarnings(), or sinon.stub(mw.log, 'warn').
But, in this case I think we can just remove it at the source as
the warning isn't helping anyone.

These types are part of the stable interface for mw.Api, and have been
for 10+ years. Remove the obsolete logging code.

Bug: T250045
Change-Id: Ibcbc32e3ad9e0afe14199d220757801bae9ad0a4