ResourceLoader: Silence "unavailable module" warning in error handling test
This is contributing to warning fatigue in CI, with other stuff
slipping and being missed because it's normal.

> Skipped unavailable module test.load.circleC
> Skipped unavailable module test.load.circle
> Skipped unavailable module test.load.missingdep

These are expected warnings for these specific error handling tests.
The test already silenced the "error" to avoid a test failure,
but it did not yet silence the warning that goes along with it.

Bug: T250045
Change-Id: I492e31f2fc2e3342742a10d9fe14a00badca5657