Establish shared component collection
The Codex Collective, a new group of WMF staff who aim to
support the Codex design system and general front-end
infrastructure, would like to move forward with the creation
of a set of shared Vue.js components in MediaWiki core. These
components are too MediaWiki-specific to be included in Codex
itself, but could be used across features. Sharing these
components will increase developer velocity and consistency of
user experience across these features.

This patch simply establishes a directory with a README
outlining how we intend for the shared components to be built
and governed.

Bug: T375220
Co-Authored-by: Anne Tomasevich <atomasevich@wikimedia.org>
Change-Id: I1e3258f6a87ae003382f8bd66a1e074178de40ef