Add prefix option to RedisConnectionPool
This adds a new prefix option to Redis to transparently prefix all
get/set operations with a specified prefix. The use case for this is
when providing a multitenant Redis/valkey cluster for multiple wikis
that are not necessarily part of the same wiki farm and thus should have
fully-isolated keyspaces.

This patch additionally adjusts the doc comment type for the password
field to align with reality. The class was already defaulting password
to null, and the Redis connection class itself accepts both a single
string as a password as well as an array with 2 elements defining the
username and password.

Change-Id: I9c39a0a60e01ecc29f76b98ed5f8e4e28802c533