Refactor RightsLogFormatter
The formatter's getMessageParameters method was tightly coupled with
the structure of the log entry. It meant that extensions providing
their own rights-like log had to reimplement a substantial portion
of this class.
Now, in such cases, extensions could override a relevant method
and leave the common code in the base RightsLogFormatter.

Bug: T369466
Change-Id: I346fb01bf44f729bb06b2101630b9b45f0c39175