Clearly show changes to user groups in rights log
Previously, the log displayed a list of former groups and a list of new
groups, which wasn't particularly helpful when user had a ton of groups.
Now, the log message is split into four parts: added groups, removed
groups, changed expiration, and groups kept intact.
This style of presentation highlights what has changed but still allows
user to reconstruct the whole set of groups before or after the change
was made (as it was possible before this patch).

Bug: T369466
Change-Id: I80d22ef7cbdc60a1437e26b6f9a6ee11bbecab64