logger: Make log() methods return void
psr/log 3.0.0 adds this return type, forcing all subclasses to have it
too. MediaWiki does not yet declare compatibility with that version, but
if people don’t use the composer merge plugin (or do use it, but also
still have separate vendor directories in some extensions or skins e.g.
to run development tools like phpcs), and they are also on PHP 8.0 or
later (on PHP 7.4, composer should not permit installation of psr/log
2.0.0 or later which requires PHP 8), then it’s possible for them to end
up with psr/log 3.0.0 in some other vendor directory, and this version
may then override the one MediaWiki depends on. In general, this kind of
composer accident is not something we can protect against (MediaWiki
can’t support all incompatible library versions that may end up in the
autoload path), but for this specific case the fix is very simple, fully
compatible with the older version of psr/log as well (adding a return
type in a subclass when the superclass doesn’t have one is always
allowed), and something we’ll have to do sooner or later anyway, so I
propose we just do it now.

Bug: T356451
Change-Id: I49562ac7f1a71e82cab79fe44296feea573e26d4